Split of Context classes
